analytics-engine_2/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
pandas==2.2.3
numpy==2.1.3
pydantic==2.9.2
python-dotenv==1.0.1
dnspython==2.7.0